[Feat]: CurrencyConverter sturct를 생성하여 scrollview에 있는 로직 분리
